refactor(api): separate /users logic
